pyproject.toml
[tool.poetry]
name = "gpt-researcher"
version = "0.8.5"
description = "GPT Researcher is an autonomous agent designed for comprehensive online research on a variety of tasks."
authors = ["Assaf Elovic <assaf.elovic@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <4"
beautifulsoup4 = ">=4.12.2"
colorama = ">=0.4.6"
duckduckgo_search = ">=4.1.1"
md2pdf = ">=1.0.1"
openai = ">=1.3.3"
python-dotenv = ">=1.0.0"
pyyaml = ">=6.0.1"
uvicorn = ">=0.24.0.post1"
pydantic = ">=2.5.1"
fastapi = ">=0.104.1"
python-multipart = ">=0.0.6"
markdown = ">=3.5.1"
langchain = "^0.3.18"
langgraph = ">=0.2.73,<0.3"
langchain_community = "^0.3.17"
langchain-openai = "^0.3.6"
tavily-python = ">=0.2.8"
permchain = ">=0.0.6"
arxiv = ">=2.0.0"
PyMuPDF = ">=1.23.6"
requests = ">=2.31.0"
jinja2 = ">=3.1.2"
aiofiles = ">=23.2.1"
SQLAlchemy = ">=2.0.28"
mistune = "^3.0.2"
htmldocx = "^0.0.6"
python-docx = "^1.1.0"
lxml = { version = ">=4.9.2", extras = ["html_clean"] }
unstructured = ">=0.13"
tiktoken = ">=0.7.0"
json-repair = "^0.29.8"
json5 = "^0.9.25"
loguru = "^0.7.2"
websockets = "^13.1"
firecrawl-py = "^1.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')"
]
asyncio_mode = "strict"
addopts = "-v"
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_fixture_loop_scope = "function"
